src/requirements.txt
requests
dnspython
pydantic
py-zabbix
